download_files now uses the standard SSL CAs.
Because most of the operations happen against the signing servers, we
create an ssl context that specifies the releng CA cert as the trusted
SSL CA.  However, we also download files to sign from other servers,
ideally over https.  This patch creates a second session that uses the
standard SSL CAs, and uses that session for `download_files`.  As long
as the servers we're downloading from use real SSL certs, we're good.
If we start needing to download from a different self-signed or
non-default CA SSL server, we'll need a slightly different solution.