use `validate_artifact_url` for validation and file paths
Previously, we were using the basename of a url download.  This works
when downloading and signing single files, or when the filenames are
unique.  It falls apart when there are multiple files with the same
filename, but different paths, e.g. the en-US apk vs the multi apk.

`validate_artifact_url` allows for us to both validate the URL is a
valid looking, well-formed taskcluster URL pointing at the artifacts of
a dependent task, and gives us the relative path of the file.  We now
use this relative path as the filepath to download, sign, and copy to
the `artifact_dir`.