bug 1390552 - move widevine .sig files to Contents/Resources/
Specify an alternate location under `Contents/Resources/` for the .sig
files, so we don't break macapp signing. Given a path with an app under
another app, we only replace the final `MacOS` with `Resources`

This also fixes the duplicate "Signing FILE with FORMAT" logs, and fixes
the compression logic for `_get_tarfile_files`.