Nixify
This adds Nix expressions to nixify shipitsriptworker. It is also
possible to generate a docker image (mostly used as an example, because
there is no config file, nor a way to pass it and run shipitscript) and
a tarball. The generated tarball can be used as a interim solution in
puppet.