Bug 1542500 Finding the right upstream artifact (#211)
* Bug 1542500 Finding the right upstream artifact

In order to translate the full local scriptworker path into an artifact, we
were using `os.path.basename` and then searching through the artifact map,
ignoring locales. This caused the first match to be returned, whether or not it
was correct.

This new variant passes in both the full local path and the upstream artifact
path, so we can match against the full artifact path instead, removing a chance
of duplicate matches.