go back to 5 attempts per notarization call
I added 10 attempts to give Apple more time to get things working on the
back end. However, today it seems to just delay when we get the error
messages, rather than resulting in a slower but green task.

If we decide we want to go back to 5 retry attempts, this is how we'd do
it.