Bug 1578497 - Notarize pkg files as well as dmg, in a single upload to Apple
Builds the pkg files immediately after app signing so they can be included in the zip uploaded to Apple. This means we end up with a notarized pkg of an un-notarized app, but Gatekeeper on Catalina is fine with this (the quarantine bit is not set on the app after install, and Gatekeeper will get the ticket for the app at some point in the background). We staple both app and pkg once Apple completes their checks.