No bug - Stop signing resources of inner apps
The missing continue instruction resulted in us logging that we'll skip signing resources but going on to sign them. We were getting files like Firefox.app/Contents/MacOS/updater.app/Contents/._PkgInfo, which has the signing stored as an extended attribute (xattr) because the file doesn't support signing natively. pkgbuild converts the xattr to the ._<file> for 'transport'. We avoided this problem for apps by setting COPYFILE_DISABLE=1 for zipping them up. The macOS package installer was probably dropping the attr, as they're not present after the install as a separate file.

In any event, avoiding this signing means the app and pkg only differ by the notarization ticket (at Firefox.app/Contents/CodeResources).