[treescript] Don't treat `push` as an action.
The main motivation is that merge day requires pushing to multiple repos,
so the logic around pushing for that needs to be distinct from the logic
for pushing from other actions.

Given that, it seems like it makes more sense to represent pushing as a boolean.
We already have something like that (i.e. `dry_run`) but that only disables pushing
if it is set, and can't enable it. This introduces a new booleam payload
attribute `push` that enables pushing if it is set (and `dry_run` is not). If it isn't
set, it falls back to looking for `push` in the actions.

In the future, the `push` action should be removed, and the `dry_run` attribute.