[beetmoverscript] fix tests: avoid circular import
script.py was importing task.py, which was importing script.py. The only
reason task was importing script was to determine `valid_actions`. We
could just drop this check, since `async_main` also ensures we exit on
an invalid action. However, to keep the behavior a bit closer to
existing, I added a `valid_actions` kwarg to `get_task_action`. If
specified, we check the action is a member of `valid_actions`.