[treescript] Set mercurial config via files. (#178)
* [treescript] Log mercurial config as part of startup.

* [treescript] Set all mercurial config via files.

This make the logging clearer, because we are not passing all the config on
every call to hg. This also properly enable the firefoxtree extension, as we
don't set `HGRCPATH` (which was overriding the system-level config file in
production).

This also switches to using the robustcheckout from the version-control-tools
checkout we have in production.