Bug 1619767 - [treescript] Do not use upstream_repo on Thunderbird clones.
As there is no comm-unified repository with all of the various branch heads,
separate pulls are needed for the to_branch and from_branch when doing merges
on those repositories.

The separate pulls are triggered by a falsy value for upstream_repo. Usually
that's set to comm-central an hg robustcheckout does what it needs to do, but
it's not required to set it.