pin k8s-image task to a dockerhub image
As noted in
https://github.com/mozilla-releng/scriptworker-scripts/pull/371#issuecomment-880254355
, the k8s-image task is the last holdout still using dind to build
docker images, rather than Kaniko.

This is problematic: the docker-image task can't pull the ancient
docker:1.6.2 image needed to build dind images. We're running up against
the 1 year expiry of these artifacts.

The real fix is to switch our k8s-image tasks to use Kaniko to build the
images, and find a way to either publish those images to docker hub (via
Kaniko? some other solution?), or switch cloudops-jenkins to download
the docker images some other way (Tom Prince had a wip solution to send
a pulse message to cloudops, and for cloudops to download from
taskcluster artifacts).

This solution, of pushing the soon-to-expire k8s-image docker image to
docker hub, and pulling from there, extends the runway.