Allow 'keyid's in intermediate authenticode certs
The upcoming Authenticode EV cert uses a different intermediate cert
from the other authenticode formats. When using a keyid,
`sign_authenticode` will now check if a ca cert with that keyid is
configured first. If none are found, it will fallback to the original
one.

This way we can support keyids in intermediate certs while preserving
backwards compatibility.