beetmoverscript: check destination path in push-to-maven
`check_maven_artifact_map` now verifies that the product name in the
task payload matches the destination paths we're told to upload to.

This should prevent uploads to the wrong place because product names
need to match entries in the product_buckets config, as a secondary
protection to the permissions granted to cloud credentials.