Create sig files for the new macOS GMP Executable (#745)
* Create sig files for the new macOS GMP Executable

On macOS, the GMP child process will transition from using the plugin-container executable to a new executable named "<App Name> Media Plugin Helper". Add the different executable names for each channel to the list of files which we need to generate .sig files for.

* Create sig files for the new macOS GMP Executable

On macOS, the GMP child process will transition from using the plugin-container executable to a new executable named "<App Name> Media Plugin Helper". Add the different executable names for each channel to the list of files which we need to generate .sig files for.

* Update comments