Bug 1853416 - [treescript] Split hg and git logic apart, r?#releng-reviewers
On the surface, this commit is silly because it is duplicating a lot of
logic that was previously refactored.

But this commit is a pre-cursor to later commits that will refactor the
Github case to use Github's GraphQL API rather than operating on a local
checkout.

By pulling them apart we accomplish a couple things:

1. It will make the follow-up commit easier to review because the
   refactoring was already done here.
2. By keeping the two use cases distinct, it will make it easier to
   remove the Mercurial bits when the time comes.

Eventually, there will only be one way of doing things again, at which
point we can move the `github` submodules back up to the top-level (if
desired).