pushapkscript: set default socket timeout to 5 minutes
The 60s default used by googleapiclient is apparently not enough for
fenix AAB uploads, so give it some buffer.

As far as I can tell, this becomes necessary because of
https://github.com/mozilla-releng/mozapkpublisher/commit/4e2ccf975e2d0fc0adfb409e892720b7e53ca0db
where I changed mozapkpublisher to not pass a fresh httplib2.Http object
and rely on googleapiclient's default `build_http` function instead,
where the timeout is set.