Bug 1876573 - Allow 'bitrisescript' to specify multiple scope prefixes
It turns out that we need to use consistent scope prefixes as the other
scriptworkers as Taskgraph configures a single prefix across all
scriptworkers. This means that the repo name can appear in the prefix,
which in turn means we'll need to support multiple valid prefixes per
trust domain.

This is how the other scriptworkers work, so it's a good change to make
anyway.