Bug 1891815 - Increase global task timeout to two hours
We have bitrisescript tasks that take longer than an hour and it looks
like we currently have no way to set a per script timeout (see the
comment about json-e in scriptworker.yml).

So the quick fix here is to bump the global timeout. I believe it's
extremely rare (if not impossible) for us to hit this timeout anyway,
because Kubernetes has its own timeout (`terminationGracePeriodSeconds`)
that is much shorter than this one. The only reason I even knew this
timeout exists is because I had set the above to `7200` for
bitrisescript. All other scripts have this value set to `3600` or less,
which means they should get force killed by k8s long before we ever
reach the scriptworker timeout.