fix(treescript): use separate Github apps for prod vs dev
Initially I didn't bother splitting these out because I figured they
both require the exact same permissions on the repo anyway. So if the
dev app secret were compromised, you'd be just as hosed as if the prod
app secret were compromised.

However, it didn't occur to me that we might only install the `dev` app
on staging repos and that we don't want a bug on treescript-dev to
accidentally expose the `prod` app's private key.

To this end, I created a `releng-treescript-dev` app. We should probably
avoid installing this app on production repos, and instead only add it
to `staging-*` repos. To help with this, I restricted the app to the
`mozilla-releng` org.