feat(beetmoverscript): Add translations trust domain
As noted in a couple of comments, this trust domain is a bit odd right now because production runs at level 1, so we don't have the prod/fake-prod distinction here that we have for other trust domains. We will be remedying this, but we're not blocking adding beetmover tasks on that.