treescript: add support for fetching l10n revisions in a github repo (#1015)
* feat(treescript): Add support for fetching head revision with github client

Mostly a refactor of existing code.

* feat(treescript): Allow git and hg configuration simultaneously

* feat(treescript): Enable git for Firefox treescript

* refactor(treescript): Only bump l10n in merge day action when an l10n_bump action is also present

This moves the decision about bumping l10n up to a higher level, which simplifies adding support for bumpning via github in the next commit.

* feat(treescript): Add support for fetching l10n revisions from github.

Most of this is copy/paste to avoid touching any of the existing l10n bumping code.

The most notable change is in build_revision_dict_github, which is vastly simplified compared to its hg compared part. In part this is because we only have one string repo to worry about, but also because @eemeli declared that we don't need to support pinning revisions anymore - so I didn't implement that here. After we cut over to the Github repo and know we aren't going back, we can rip out the hg support.

Most of the rest of this is just wiring in calls to l10n_bump_github when appropriate.

* force treestatus check off for testing

* style: pre-commit.ci auto fixes [...]

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>