Don't retrieve logs if a job gets cancelled
Note that while it would look better to have an early return in the
`finally`, we can't do that as returning from `finally` eats the
exception and it doesn't get raised to the parent (which in this case
would be in charge of actually cancelling the build).

Skipping the log retrieval here is necessary because it would just wait
for the build to finish again, defeating the point of cancelling the
job.