signingscript: import new authenticode certs
These new certs are associated with the new authenticode keys in the Autograph GCP production environment, so we need new formats to along with them. Both the EV and non-EV cert are using same intermediates as before, so we don't need any changes to the CA bundles.

I've slightly simplified the `sign.py` blocks that deal with certs and cafiles as a part of this, which I'm _fairly_ certain is safe, but I'll be testing in dev to gain some more confidence in that.