signingscript: autograph gcp migration step 4: add gcp_prod entries for L3 workers (#1123)
This adds `gcp_prod` prefixes for everything in the production workers, except for the authenticode stub format in `vpn`, which it doesn't actually use.

The new authenticode keyids require new hawk credentials. There is a `gcp-prod-level-3` branch in the sops repo with the necessary changes staged in it that needs to be pushed before this rolls out.