move GithubClient to scriptworker_client; add support for slashes in key names
A few things going on here, mostly boring:
* Move the existing GithubClient to scriptworker_client
* Move the associated test fixtures to a new `pytest-scriptworker-client` package
* Necessary adjustments in treescript to let it be able to find and use the new packages

Although not ideal, this also includes a minor fix to allow slashes in GraphQL key names that I had already incorporated into the version of this code in https://github.com/mozilla-releng/scriptworker-scripts/pull/1135. I can pull it out if necessary, but I think it's probably small and safe enough to just keep in here?