fix(scriptworker_client): gracefully handle missing files in github_client.get_files
At the moment we end up throwing an exception if a file is missing, which means we don't get the contents of _any_ of the files fetched, which is probably not what we want.

This patch changes things such that missing files will still have a key in the returned dict, but with a None value. (This allows for distinguishing between empty files and missing files.)