fix(landoscript): rename `central` branch references to `main`
The `central` name is going away as part of the Github migration. What is now `mozilla-central` will be known as `firefox-main` in the Lando API, and will be associated with the `main` branch of the new Github repository.

I expect I will rename the `behaviours` in the Gecko repo as well, so I've renamed the references to `central_to_beta` and `bump_central` in tests here as well.