add landoscript (#1135)
* feat(landoscript): initial landoscript code with support for version bump action

This adds the rough structure for landoscript as well as implementing the `version_bump` action (necessary to make it practical to test the initial code).

* feat(landoscript): implement support for `tag` action

* refactor(landoscript): add tests that run multiple actions in one run

Most notably, this moves common set-up into conftest, and tests that aren't testing action-specific logic (eg: lando submission) into test_script.

* feat(landoscript): implement `l10n_bump` action

The helper functions here are copied out of treescript (which will soon be EOL'ed).

Also included here is some minor refactoring to avoid duplication of common of `create-commit` logic.

* feat(landoscript): implement `merge_day` action

Most of the helpers here are copied out of treescript, with some tweaks and simplications where it was possible (mostly to get rid of now-unnecessary logic).

Some refactoring of other actions was done here as well, to make it possible to call them from the `merge_day` action. Most notably: the `version_bump` action has been updated to support multiple version bumps in one run, which allows us to do all the merge day version bumps in a single commit, as we do now with treescript.

Additional test refactoring/movement was also done to make many of the helpers available to merge day tests.

* refactor(landoscript): create `run_test` helper that simplifies most of the happy path tests

This new helper allows a huge simplication for most of the action-specific tests.

* feat(landoscript): implement android l10n actions

The implementation of this is a fairly significant departure from the treescript one in two ways:
* It supports these actions without having a Gecko or `android-l10n` tree available. This necessitated fetching the remote file listings of these repositories, and using `moz.l10n` instead of `compare-locales`.
* I've fully separated out implementation of the actions. Although at a very high level they look similar, the details are different enough that IMO it's much easier and better to duplicate some of the code rather than add the indirection to avoid it. (I found it very difficult to read the treescript implementation because of this, and I didn't want to do the same here).

A necessary part of doing this was some enhancements to `diff_contents` to support added and removed files properly.

* fix(landoscript): don't support dontbuild/ignore closed tree configuration outside of l10n bump

As it turns out, these are static for all other types of actions.

* feat(landoscript): add authorization header when submitting lando requests

Also ensure that LANDO_API and LANDO_TOKEN are set up during startup.

* fix(landoscript): address review comments around lando api requirements

* feat(landoscript): pull repository url and branch from lando instead of config

* fix(landoscript): Replace TypedDicts with dataclasses

This required a little bit of massaging for places where there's nested dataclasses, but it was otherwise straightforward.

This change also uncovered some places where tests were including unnecessary data, which have been fixed. It also replaced some existing null checks (because the dataclass will throw if a required member is not present during construction).

* fix(landoscript): improve comment around some test code