fix(landoscript): improve esr version bump tests and verification
This adjusts the `version-bump` test for ESR to cover the real world scenario of bumping the esr-suffixed and non-suffixed version files at the same time - something I uncovered an issue with during some local testing in dryrun mode.

I also noticed that the verifications done for version bumps were masking problems when an extra suffix was included on the last line of a file (because we weren't matching the newline character, extra trailing characters would not cause the verification to fail).