fix(landoscript): ensure esr bumps use `esr` suffix where appropriate
While inspecting some diffs that were spat out in dry run mode, I noticed that `browser/config/version.txt` ended up with an `esr` suffix on it when running `release-to-esr` merge-automation for esr128. As it turns out, this was happening in tests as well, but masked by an incomplete assertion (now fixed).

The fix I've implementing here is not great: it's a flag that only really affects anything for the `release-to-esr` merge automation case, but it's the best I could come up with for now. One would think it would be preferable to drop the esr logic in `get_cur_and_next_version` - which indeed fixes this problem, but it it ends up causing issues with the regular `version-bump` on esr: we end up with an esr-suffixed version in `version.txt`, because there's no per-file granularity in the `version_bump` action -- all files take the same version number. (To be quite honest - I think this is a problem in treescript as well, and it's avoided because we always end up in the "Version bumping skipped due to conflicting values" code path: https://github.com/mozilla-releng/scriptworker-scripts/blob/419abe11d20ca5c5a123aaaa5d010f428dff9ca4/treescript/src/treescript/gecko/versionmanip.py#L164C26-L164C75, and end up bumping the version through the `bump-esr` merge automation instead -- but I didn't want to perpepuate the problem here, in case it ends up mattering in the future.)