fix(scriptworker_client): make prefix correct for recursive get_file_listings queries (#1174)
* fix(scriptworker_client): make prefix correct for recursive get_file_listings queries (#1170)

While doing some more real world testing of this code against GitHub I realized that the tests for this feature return the wrong responses when a `path` is given: they should not include leading directories in the response, only the children of the `path` given. Fixing this actually cleaned up the real code as well: as long as we path `prefix` to `_process_file_listings`, the files returned will always have the correct path, and we don't need to do any munging for the `refetches`.

* fix(landoscript): adjust landoscript tests after changes to `get_file_listing`

* style: pre-commit.ci auto fixes [...]

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>