feat(landoscript): respect closed tree for android l10n sync (#1163)
* feat(landoscript): respect closed tree for android l10n sync

I originally thought this was only used for l10n-bump, but it turns out we need to check treestatus for android l10n sync tasks as well.

I also added a test for the "dontbuild is only respected for l10n-bump" codepath while I'm touching this area.

* fix(landoscript): test treestatus for both open and approval required status

I've switched to "approval required" as the default state (because that's what we typically see for beta, where it's actually checked), while maintaining a test for the "open" state as well.