feat(landoscript): support target revisions in tagging, and correct tag action payload
When I originally implemented the `tag` action I failed to use the right payload. This patch addresses that, including adding support for target revisions.

Of note with the latter is that for direct tagging (eg: `release-early-tagging`) we need to accept an hg revision (to avoid taskgraph needing to look up the git revision). This revision + repo url can be used to look up the git revision.

For merge day tagging, the revision is not known up front: we use whatever the tip of the to and/or from branches, which we can look-up directly in github.

When we move decision tasks over to firing from Github events, we can drop the hg support.