fix(landoscript): a set of non-urgent follow-ups (#1200)
* fix(landoscript): drop unused hg_mozilla_uri from schema

* fix(landoscript): drop unused touch_clobber_file parameter: it's always True

* fix(landoscript): improve lando action printing

* don't print out actions twice in non-dry run mode
* don't print out diff contents in dry-run mode

* feat(landoscript): write out lando actions to a task artifact

Just a nice to have, to make interpreting task output slightly easier. It also allows us to have the full versions of `create-commit` actions now that we're omitting parts of them for brevity in the log.

* fix(landoscript): remove/update stray TODOs

* Mapping hg -> git revisions is sorted; drop the old comment in the tests.
* Update the comment in the tag action to reflect that we're failing if we can't map the revisions, but that it doesn't work well for Try testing at the moment.

* fix(landoscript): remove guard around 'git_commit' in tag action

We always require a revision to tag, and we always have one in both code paths, unless we error out.

* fix(landoscript): make `version_files` optional for merge day

This was required once upon a time, but became optional when `early-to-late-beta` was implemented.