fix(landoscript): use bump branch for replacements and clobber file updates
This ought to have been done as part of https://github.com/mozilla-releng/scriptworker-scripts/pull/1175, where a similar adjustment was made for version bumps. Whenever we're making changes to something post-merge, we need to use the `from` version of the file as the original. Failing to do so will generate incorrect diffs that fail to apply if the contents differ between the `from` and `to` branches.