Bug 1946095 - Add support for the samsung galaxy store to pushapkscript
The target store was already part of the task schema as a leftover from
amazon. I just added support for samsung there. Because it also has
support for rollouts, I made that non google specific. I also overloaded
the `commit` parameter although a dry run on the samsung galaxy store
won't do anything beside checking that the arguments passed in and the
APKs are correct.