bug 1967787: feat: handle landoscript terminations better (#1217)
* add clarifying comment to pre-stop.sh

* refactor(landoscript): move action processing to its own function

This will make the subsequent work of skipping it we find a lando job in a previous run easier.

* bug 1967787: feat: handle landoscript terminations better

This improves landoscript reruns by looking for the lando status URL from an earlier run, and polling it instead of redoing work (that is likely to fail). I don't think we can be _perfect_ here, but this is much better than the current state of affairs.