feat(beetmover): add support for uploading payload data as a file (#1222)
The motivation here is to make uploading the experiment configuration for translations easier (see https://github.com/mozilla/translations/issues/466). This config comes in as action task input, and is never published as a task artifact in the form that we want it. Rather than create an unnecessary file (most likely through a special task that simply creates that file out of parameters), let's add support for writing out files that come into beetmover as data.

I'm not _super_ happy about requiring base64 encoded data, but I'm also not sure of a better alternative. (At one point I allowed JSON objects to given instead of a base64 encoded string - but that only serves a much narrower set of use cases.)