fix(landoscript): ensure a newline exists between diff sections
I believe this broke with https://github.com/mozilla-releng/scriptworker-scripts/pull/1203, where `diff_contents` (correctly) stopped always having a newline character at the end of a diff, but callers were not updated to handle it. This wasn't a problem in production because AFAICT, the only file we ever touch in landoscript that doesn't have an EOF newline is `CLOBBER`, which happens (luckily) to be the last file in the `merge_day` diff.

Unfortunately, this was masked by a bug in the tests where we split on `diff\n` instead of `\ndiff` - resulting in the diffed files not being split correctly.