Fixup docker images used for testing
For some reason there was a python and a base-test image doing basically
the same thing except one was able to be used as a base to install more
packages. This merges both into a single base-test image.

I haven't looked into this but I can only guess that the python image
used to be the base image for both tests and prod, but that the prod
one diverged to the now base image