Traverse the github pushlogs in the right direction when looking for a shippable revision (#1239)
I completetely missed that when I wrote this but github returns commits
from old to new and we need to traverse them in reverse order to avoid
trying to ship the latest available candidate.

The test I added here is based on real data that the github API returned
for the one release where we detected that.