Bug 1975771 - Make sure scriptworkers get the memo about having to stop
If k8s tells a worker to stop before the worker has setup its signal
handlers, said worker would never receive the USR1 one and continue
picking jobs despite being on a timer. If a task is still running when
the timer expires, pre-stop would exit and k8s would end up sending a
sigterm to the task which results in `Automation Error: python exited
with signal -15`.

This changes the pre-stop behavior to continue sending the signal while
it's waiting. This is safe to do as the receiving side is idempotent.