Respect the uv.lock file when installing packages in docker container
This works by copying the workspace definition and the uv.lock file
within the container and then syncing only the package that is
interesting to us ($SCRIPT_NAME). This also gets rid of the manual
installation of scriptworker_client which is now unnecessary and
should've been a red flag in the first place but I completely missed it,
it will now only be installed when a package requires it, and because it
is contained in the workspace, it will just work (tm). We could probably
avoid copying scriptworker_client in the container for most scripts but
the gain is not worth splitting the image definition between scripts
with and without the client lib installed.

Fixes #1251