fix: don't return earlier lando status for merge conflicts
Merge conflicts end up resolving a run with the `intermittent-task` exception. We want to keep using this exception, because automatic retries are a feature here. However, we need to make sure that when we resume from this run that we start fresh and ignore the earlier lando attempt. To do this, we need to enrich the status artifact we add with a distinct failure reason.

Fixes https://bugzilla.mozilla.org/show_bug.cgi?id=1978590